Pull Latest changes from main project

+++ docker/build_standalone-tcks.sh
@@ -38,6 +38,11 @@ echo "########## Trunk.Install.V5 Config ##########"
 wget --progress=bar:force --no-cache $GF_BUNDLE_URL -O latest-glassfish.zip
 unzip -o latest-glassfish.zip
 
+if [ ! -z "$GF_VERSION_URL" ]; then
+  wget --progress=bar:force --no-cache $GF_VERSION_URL -O glassfish.version
+  cat glassfish.version
+fi
+
 export ANT_OPTS="-Xmx2G -Djava.endorsed.dirs=${BASEDIR}/glassfish5/glassfish/modules/endorsed \
 -Djavax.xml.accessExternalStylesheet=all \
 -Djavax.xml.accessExternalSchema=all \
@@ -64,65 +69,118 @@ fi
 ### STANDALONE TCK BUNDLE BUILD - START
 ################################################
 
-DOC_SPECIFIC_PROPS=" "
-JAXWS_SPECIFIC_PROPS=" "
+DOC_SPECIFIC_PROPS=""
+JAXWS_SPECIFIC_PROPS=""
+
+#Generate Version file
+GIT_HASH=`git rev-parse HEAD`
+GIT_BRANCH=`git branch | awk '{print $2}'`
+BUILD_DATE=`date`
+echo "Git Revision: ${GIT_HASH}" >> ${WORKSPACE}/version.info
+echo "Git Branch: ${GIT_BRANCH}" >> ${WORKSPACE}/version.info
+echo "Build Date: ${BUILD_DATE}" >> ${WORKSPACE}/version.info
+
+
+if [ ! -z "$TCK_BUNDLE_BASE_URL" ]; then
+  mkdir -p ${WORKSPACE}/standalone-bundles/
+  for tck in ${TCK_LIST[@]}; do
+    echo "Skipping build and using pre-build binary TCK bundle: ${TCK_BUNDLE_BASE_URL}/${tck}tck.zip"
+    wget  --progress=bar:force --no-cache ${TCK_BUNDLE_BASE_URL}/${tck}tck.zip -O ${WORKSPACE}/standalone-bundles/${tck}tck_latest.zip
+  done
+  exit 0
+fi
 
 for tck in ${TCK_LIST[@]}; do
   if [ "jaxr" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Djwsdp.home=$GF_HOME/glassfish5/glassfish/ -Dts.classpath=$GF_HOME/glassfish5/glassfish/modules/endorsed/jaxb-api.jar:$GF_HOME/glassfish5/glassfish/modules/endorsed/webservices-api-osgi.jar:$GF_HOME/glassfish5/glassfish/modules/javax.xml.registry-api.jar:$GF_HOME/glassfish5/glassfish/modules/webservices-osgi.jar:$GF_HOME/glassfish5/glassfish/modules/jaxb-osgi.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar:$BASEDIR/lib/javatest.jar:$BASEDIR/lib/tsharness.jar -Dall.test.dir=com/sun/ts/tests/jaxr,com/sun/ts/tests/signaturetest/jaxr -Dbuild.level=2"
+    TCK_SPECIFIC_PROPS="-Djwsdp.home=$GF_HOME/glassfish5/glassfish/ -Dts.classpath=$GF_HOME/glassfish5/glassfish/modules/endorsed/jakarta.xml.bind-api.jar:$GF_HOME/glassfish5/glassfish/modules/endorsed/webservices-api-osgi.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.xml.registry-api.jar:$GF_HOME/glassfish5/glassfish/modules/webservices-osgi.jar:$GF_HOME/glassfish5/glassfish/modules/jaxb-osgi.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar:$BASEDIR/lib/javatest.jar:$BASEDIR/lib/tsharness.jar -Dall.test.dir=com/sun/ts/tests/jaxr,com/sun/ts/tests/signaturetest/jaxr -Dbuild.level=2"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jaxrpc" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Dendorsed.dirs=$GF_HOME/glassfish5/glassfish/modules/endorsed -Dall.test.dir=com/sun/ts/tests/jaxrpc,com/sun/ts/tests/signaturetest/jaxrpc -Dbuild.level=2 -Dlocal.classes=$GF_HOME/glassfish5/glassfish/modules/webservices-osgi.jar:$GF_HOME/glassfish5/glassfish/modules/javax.xml.rpc-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.mail.jar:$GF_HOME/glassfish5/glassfish/modules/jaxb-osgi.jar:$GF_HOME/modules/javax.ejb-api.jar"
+    TCK_SPECIFIC_PROPS="-Dendorsed.dirs=$GF_HOME/glassfish5/glassfish/modules/endorsed -Dall.test.dir=com/sun/ts/tests/jaxrpc,com/sun/ts/tests/signaturetest/jaxrpc -Dbuild.level=2 -Dlocal.classes=$GF_HOME/glassfish5/glassfish/modules/webservices-osgi.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.xml.rpc-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.mail.jar:$GF_HOME/glassfish5/glassfish/modules/jaxb-osgi.jar:$GF_HOME/modules/jakarta.ejb-api.jar"
     DOC_SPECIFIC_PROPS="-propertyfile $BASEDIR/install/jaxrpc/bin/build.properties -Dts.home=$BASEDIR -Ddeliverable.version=1.1 -Ddeliverable.class=com.sun.ts.lib.deliverable.jaxrpc.JAXRPCDeliverable"
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jta" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Djta.classes=$GF_HOME/glassfish5/glassfish/modules/javax.transaction-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar"
+    TCK_SPECIFIC_PROPS="-Djta.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.transaction-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jsf" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Djsf.classes=$GF_HOME/glassfish5/glassfish/modules/cdi-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet.jsp.jstl-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.inject.jar:$GF_HOME/glassfish5/glassfish/modules/javax.faces.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet.jsp-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.el.jar"
+    TCK_SPECIFIC_PROPS="-Djsf.classes=$GF_HOME/glassfish5/glassfish/modules/cdi-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet.jsp.jstl-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.inject.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.faces.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet.jsp-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.el.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jsonp" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Djsonp.classes=$GF_HOME/glassfish5/glassfish/modules/javax.json-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.json.bind-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.json.jar"
+    TCK_SPECIFIC_PROPS="-Djsonp.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.json.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.json.bind-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.json.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jsonb" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Djsonb.classes=$GF_HOME/glassfish5/glassfish/modules/javax.json-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.json.bind-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.json.jar:$GF_HOME/glassfish5/glassfish/modules/javax.json.jar:$GF_HOME/glassfish5/glassfish/modules/javax.inject.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar"
+    TCK_SPECIFIC_PROPS="-Djsonb.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.json.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.json.bind-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.json.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.json.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.inject.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jaxrs" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Djaxrs.classes=$GF_HOME/glassfish5/glassfish/modules/javax.json-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.json.bind-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.json.jar:$GF_HOME/glassfish5/glassfish/modules/jsonp-jaxrs.jar:$GF_HOME/glassfish5/glassfish/modules/javax.ws.rs-api.jar:$GF_HOME/glassfish5/glassfish/modules/jsonp-jaxrs.jar:$GF_HOME/glassfish5/glassfish/modules/endorsed/javax.annotation-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.json.jar:$GF_HOME/glassfish5/glassfish/modules/javax.ejb-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.interceptor-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/cdi-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.inject.jar:$GF_HOME/glassfish5/glassfish/modules/validation-api.jar:$GF_HOME/glassfish5/glassfish/modules/bean-validator.jar"
+    TCK_SPECIFIC_PROPS="-Djaxrs.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.json.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.json.bind-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.json.jar:$GF_HOME/glassfish5/glassfish/modules/jsonp-jaxrs.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.ws.rs-api.jar:$GF_HOME/glassfish5/glassfish/modules/jsonp-jaxrs.jar:$GF_HOME/glassfish5/glassfish/modules/endorsed/jakarta.annotation-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.json.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.ejb-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.interceptor-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/cdi-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.inject.jar:$GF_HOME/glassfish5/glassfish/modules/validation-api.jar:$GF_HOME/glassfish5/glassfish/modules/bean-validator.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "websocket" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Dwebsocket.classes=$GF_HOME/glassfish5/glassfish/modules/javax.websocket-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.inject.jar:$GF_HOME/glassfish5/glassfish/modules/cdi-api.jar"
+    TCK_SPECIFIC_PROPS="-Dwebsocket.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.websocket-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.inject.jar:$GF_HOME/glassfish5/glassfish/modules/cdi-api.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "securityapi" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Dsecurityapi.classes=$GF_HOME/glassfish5/glassfish/modules/endorsed/javax.annotation-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.inject.jar:$GF_HOME/glassfish5/glassfish/modules/javax.security.enterprise-api.jar:$GF_HOME/glassfish5/glassfish/modules/cdi-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.faces.jar:$GF_HOME/glassfish5/glassfish/modules/javax.security.auth.message-api.jar:$BASEDIR/glassfish5/glassfish/modules/javax.ejb-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.interceptor-api.jar"
+    TCK_SPECIFIC_PROPS="-Dsecurityapi.classes=$GF_HOME/glassfish5/glassfish/modules/endorsed/jakarta.annotation-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.inject.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.security.enterprise-api.jar:$GF_HOME/glassfish5/glassfish/modules/cdi-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.faces.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.security.auth.message-api.jar:$BASEDIR/glassfish5/glassfish/modules/jakarta.ejb-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.interceptor-api.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "el" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Del.classes=$GF_HOME/glassfish5/glassfish/modules/javax.el.jar"
+    TCK_SPECIFIC_PROPS="-Del.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.el.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "concurrency" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Dconcurrency.classes=$GF_HOME/glassfish5/glassfish/modules/javax.enterprise.concurrent-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.ejb-api.jar:$GF_HOME/glassfish5/glassfish/modules/jta.jar:$GF_HOME/glassfish5/glassfish/modules/javax.enterprise.deploy-api.jar"
+    TCK_SPECIFIC_PROPS="-Dconcurrency.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.enterprise.concurrent-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.ejb-api.jar:$GF_HOME/glassfish5/glassfish/modules/jta.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.enterprise.deploy-api.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "connector" == "$tck" ]
   then
     TCK_SPECIFIC_PROPS="-Dconnector.home=$GF_HOME/glassfish5/glassfish/"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jacc" == "$tck" ]
   then
     TCK_SPECIFIC_PROPS="-Djacc.home=$GF_HOME/glassfish5/glassfish/"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jaspic" == "$tck" ]
   then
     TCK_SPECIFIC_PROPS="-Djaspic.home=$GF_HOME/glassfish5/glassfish/"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "caj" == "$tck" ]
   then
     TCK_SPECIFIC_PROPS=""
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jms" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Djms.classes=$GF_HOME/glassfish5/glassfish/modules/javax.jms-api.jar"
+    TCK_SPECIFIC_PROPS="-Djms.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.jms-api.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jsp" == "$tck" ]
   then
     TCK_SPECIFIC_PROPS="-DwebServerHome=$GF_HOME/glassfish5/glassfish/"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jstl" == "$tck" ]
   then
     TCK_SPECIFIC_PROPS="-DwebServerHome=$GF_HOME/glassfish5/glassfish/"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "jaxws" == "$tck" ]
   then
      PROXY_HOST=`echo ${http_proxy} | cut -d: -f2 | sed -e 's/\/\///g'`
@@ -133,15 +191,22 @@ for tck in ${TCK_LIST[@]}; do
     cat $BASEDIR/install/jaxws/bin/ts.jte | grep wsimport.jvmargs
     TCK_SPECIFIC_PROPS="-Dwebcontainer.home=$BASEDIR/glassfish5/glassfish -Dwebcontainer.home.ri=$BASEDIR/glassfish5/glassfish -Ddeliverable.version=2.3"
     JAXWS_SPECIFIC_PROPS="-Dwebcontainer.home=$BASEDIR/glassfish5/glassfish -Dwebcontainer.home.ri=$BASEDIR/glassfish5/glassfish -Ddeliverable.version=2.3"
+    DOC_SPECIFIC_PROPS=""
   elif [ "jpa" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Djpa.classes=$GF_HOME/glassfish5/glassfish/modules/javax.persistence.jar"
+    TCK_SPECIFIC_PROPS="-Djpa.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.persistence.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "saaj" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Dlocal.classes=$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.ejb-api.jar -Dwebcontainer.home=$GF_HOME/glassfish5/glassfish -Dendorsed.dirs=$GF_HOME/glassfish5/glassfish/modules/endorsed"
+    TCK_SPECIFIC_PROPS="-Dlocal.classes=$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.ejb-api.jar -Dwebcontainer.home=$GF_HOME/glassfish5/glassfish -Dendorsed.dirs=$GF_HOME/glassfish5/glassfish/modules/endorsed"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   elif [ "servlet" == "$tck" ]
   then
-    TCK_SPECIFIC_PROPS="-Dservlet.classes=$GF_HOME/glassfish5/glassfish/modules/endorsed/javax.annotation-api.jar:$GF_HOME/glassfish5/glassfish/modules/javax.servlet-api.jar"
+    TCK_SPECIFIC_PROPS="-Dservlet.classes=$GF_HOME/glassfish5/glassfish/modules/endorsed/jakarta.annotation-api.jar:$GF_HOME/glassfish5/glassfish/modules/jakarta.servlet-api.jar"
+    DOC_SPECIFIC_PROPS=""
+    JAXWS_SPECIFIC_PROPS=""
   else
     echo "TCK is not in the list"
   fi
@@ -152,7 +217,10 @@ for tck in ${TCK_LIST[@]}; do
   if [ "jaxrpc" == "$tck" ]
   then
     echo "Generating JAXRPC specific classes using wscompile"
-    cp $BASEDIR/lib/ant-props.jar $ANT_HOME/lib
+    # Copy additional ant library required for nested property handling in jaxrpc build
+    if [[ -f $BASEDIR/lib/ant-props.jar && ! -f $ANT_HOME/lib/ant-props.jar ]]; then
+      cp $BASEDIR/lib/ant-props.jar $ANT_HOME/lib
+    fi
     sed -i "s#webserver\.home=.*#webserver.home=$GF_HOME/glassfish5/glassfish#g" $BASEDIR/install/$tck/bin/build.properties
     sed -i "s#jaxrpc\.tool=.*#jaxrpc.tool=$GF_HOME/glassfish5/glassfish/bin/wscompile#g" $BASEDIR/install/$tck/bin/build.properties
     cat $BASEDIR/install/$tck/bin/build.properties
@@ -176,5 +244,8 @@ for tck in ${TCK_LIST[@]}; do
     strippedEntry=`echo "$entry" | cut -d_ -f1`
     echo "copying ${WORKSPACE}/release/${UPPER_TCK}_BUILD/latest/$entry to ${WORKSPACE}/standalone-bundles/${strippedEntry}_latest.zip"
     cp ${WORKSPACE}/release/${UPPER_TCK}_BUILD/latest/$entry ${WORKSPACE}/standalone-bundles/${strippedEntry}_latest.zip
+    cp ${WORKSPACE}/version.info ${WORKSPACE}/${strippedEntry}.version
   done
 done
+
+
